backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
redis==4.6.0
pydantic==2.3.0
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alpaca-trade-api==3.0.2
ib_insync==0.9.86
tensorflow==2.13.0
scikit-learn==1.3.0
nltk==3.8.1
pandas==2.1.0
numpy==1.24.3
plotly==5.16.1
requests==2.31.0
aiohttp==3.8.5
websockets==11.0.3
jinja2==3.1.2
python-dotenv==1.0.0
pytest==7.4.0
yfinance==0.2.28
beautifulsoup4==4.12.2
tweepy==4.14.0
schedule==1.2.0
hummingbot==1.10.0
aioredis==2.0.1
email-validator==2.0.0
httpx==0.24.1
python-dateutil==2.8.2
